+++ doc/conf.py
@@ -7,9 +7,15 @@ import os.path as osp
 import shutil
 import sys
 
+import guidata.config as gcfg
+
 sys.path.insert(0, os.path.abspath(".."))
 
 import sigima
+
+# Turn off validation of guidata config
+# (documentation build is not the right place for validation)
+gcfg.set_validation_mode(gcfg.ValidationMode.DISABLED)
 
 
 def copy_changelog(app):