+++ CHANGELOG.md
@@ -10,6 +10,10 @@ See DataLab [roadmap page](https://datalab-platform.com/en/contributing/roadmap.
 * New ROI features:
   * Improved single ROI title handling, using default title based on the index of the ROI when no title is provided.
   * Added `combine_with` method to ROI objects (`SignalROI` and `ImageROI`) to return a new ROI that combines the current ROI with another one (union) and handling duplicate ROIs.
+  * Image ROI coordinates:
+    * Before this change, image ROI coordinates were defined using indices by default.
+    * Now, `ROI2DParam` uses physical coordinates by default.
+    * Note that ROI may still be defined using indices instead (using `create_image_roi` function).
 
 * New image processing features:
   * New "Gaussian frequency filter" feature: